event wrapper wins the grammy for best wrap album
its not actually a wrapper
